add exemple mvc-resteasy-hello
