+++ examples/mvc-resteasy/README.adoc
@@ -13,7 +13,7 @@ Build and start the demo:
 
 Open:
 
-    http://localhost:8080/mvc-resteasy
+    http://localhost:8080/mvc-resteasy/app/hello?name=TomEE
 
 
 ==== Intro of Eclipse Krazo
