+++ examples/mvc-resteasy/pom.xml
@@ -34,8 +34,6 @@
     <tomee.version>8.0.0-SNAPSHOT</tomee.version>
     <version.ozark>1.0.0-m04</version.ozark>
     <resteasy.version>3.6.1.Final</resteasy.version>
-    <version.arquillian>1.1.13.Final</version.arquillian>
-    <version.graphene.webdriver>2.3.1</version.graphene.webdriver>
   </properties>
 
   <build>
@@ -79,13 +77,6 @@
       <groupId>org.mvc-spec.ozark</groupId>
       <artifactId>ozark-resteasy</artifactId>
       <version>${version.ozark}</version>
-      <scope>runtime</scope>
-    </dependency>
-    
-    <dependency>
-        <groupId>javax.mvc</groupId>
-        <artifactId>javax.mvc-api</artifactId>
-        <version>1.0-pfd</version>
     </dependency>
 
     <!-- RestEasy -->
@@ -104,44 +95,21 @@
       <artifactId>resteasy-servlet-initializer</artifactId>
       <version>${resteasy.version}</version>
     </dependency>
-
-     <dependency>
-      <groupId>org.jboss.arquillian.junit</groupId>
-      <artifactId>arquillian-junit-container</artifactId>
-      <version>${version.arquillian}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.tomee</groupId>
-      <artifactId>arquillian-tomee-remote</artifactId>
-      <version>${tomee.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.tomee</groupId>
-      <artifactId>apache-tomee</artifactId>
-      <version>${tomee.version}</version>
-      <classifier>webprofile</classifier>
-      <scope>test</scope>
-      <type>zip</type>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.arquillian.graphene</groupId>
-      <artifactId>graphene-webdriver</artifactId>
-      <version>${version.graphene.webdriver}</version>
-      <type>pom</type>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.11</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>net.sourceforge.htmlunit</groupId>
-      <artifactId>htmlunit</artifactId>
-      <version>2.32</version>
-      <scope>test</scope>
-   </dependency>
   </dependencies>
+
+  <!--
+  This section allows you to configure where to publish libraries for sharing.
+  It is not required and may be deleted.  For more information see:
+  http://maven.apache.org/plugins/maven-deploy-plugin/
+  -->
+  <distributionManagement>
+    <repository>
+      <id>localhost</id>
+      <url>file://${basedir}/target/repo/</url>
+    </repository>
+    <snapshotRepository>
+      <id>localhost</id>
+      <url>file://${basedir}/target/snapshot-repo/</url>
+    </snapshotRepository>
+  </distributionManagement>
 </project>
