--- examples/mvc-resteasy/src/main/java/org/superbiz/application/MVCApplication.java
@@ -1,6 +1,0 @@
-package org.superbiz.application;
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-@ApplicationPath("app")
-public class MVCApplication extends Application { }
